Refactor: SituationType 패키지 이동
