+++ src/main/java/com/boardgo/domain/boardgame/service/HomeBoardGameQueryUseCase.java
@@ -1,6 +1,6 @@
 package com.boardgo.domain.boardgame.service;
 
-import com.boardgo.domain.boardgame.entity.SituationType;
+import com.boardgo.domain.boardgame.entity.enums.SituationType;
 import com.boardgo.domain.boardgame.service.response.CumulativePopularityResponse;
 import com.boardgo.domain.boardgame.service.response.SituationBoardGameResponse;
 import java.util.List;
